backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow==2.13.0  # or compatible version for your model
